fix: warnings under Windows+MSVC+x86
Remove any remaining warnings under Windows+MSVC+x86, now the examples,
tests and benchmarks compile cleanly in 32-bit mode. I think we are
actually using the correct types, not just silencing warnings with
`static_cast<>` too.

To enable `/WX` (treat warnings as errors) I had to disable warnings
around an Abseil include, there is a TODO for a better fix.

+++ google/cloud/spanner/benchmarks/benchmarks_config.h
@@ -42,8 +42,8 @@ struct Config {
   int minimum_clients = 1;
   int maximum_clients = 1;
 
-  std::int64_t table_size = 1000 * 1000L;
-  std::int64_t query_size = 1000;
+  std::int32_t table_size = 1000 * 1000L;
+  std::int32_t query_size = 1000;
 
   bool use_only_clients = false;
   bool use_only_stubs = false;
